aider: Update CSS to make child elements stretch to fill the screen.
# Aider chat conversation:

USER: How do we update the CSS of the child elements?

ASSISTANT: To make the child elements stretch to fill the screen, we need to modify their CSS. We can use the `flex`, `flex-grow`, and `items-stretch` classes to achieve this.

The `flex` class will make the div a flex container, allowing its children to grow and shrink within it. The `flex-grow` class will make the div take up any remaining space in the flex container. The `items-stretch` class will make the children of the flex container stretch to fill the container along the cross axis.

Here are the modifications:

```tsx
app\pricing\layout.tsx
<<<<<<< ORIGINAL
    <div className="inline-block max-w-lg text-center justify-center flex-grow">
=======
    <div className="flex flex-grow items-stretch inline-block max-w-lg text-center justify-center">
>>>>>>> UPDATED
```

```tsx
app\pricing\page.tsx
<<<<<<< ORIGINAL
    <div className="container mx-auto my-8 flex-grow">
=======
    <div className="flex flex-grow items-stretch container mx-auto my-8">
>>>>>>> UPDATED
```

Please note that these changes will make the child elements stretch to fill the screen, but they won't necessarily make the content take up all space between the header navbar menu and the footer. If you want the content to take up all space between the header navbar menu and the footer, you might need to adjust the CSS of the parent elements as well.